requirements.txt
customtkinter==0.3
pillow==9.5.0
python-docx==1.1.0
screeninfo==0.8.1
ttkbootstrap==1.10.1